Rename date ranges data classes
