+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/ranges/data/LastWeekRangeData.kt
@@ -1,5 +1,6 @@
 package com.woocommerce.android.ui.analytics.ranges.data
 
+import com.woocommerce.android.extensions.endOfCurrentWeek
 import com.woocommerce.android.extensions.oneWeekAgo
 import com.woocommerce.android.extensions.startOfCurrentWeek
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRange
@@ -7,7 +8,7 @@ import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRangeData
 import java.util.Calendar
 import java.util.Date
 
-class AnalyticsHubWeekToDateRangeData(
+class LastWeekRangeData(
     referenceDate: Date,
     calendar: Calendar
 ) : AnalyticsHubTimeRangeData {
@@ -15,19 +16,18 @@ class AnalyticsHubWeekToDateRangeData(
     override val previousRange: AnalyticsHubTimeRange
 
     init {
-        calendar.time = referenceDate
-        val startOfCurrentWeek = calendar.startOfCurrentWeek()
-        currentRange = AnalyticsHubTimeRange(
-            start = startOfCurrentWeek,
-            end = referenceDate
-        )
-
         val oneWeekAgo = referenceDate.oneWeekAgo()
         calendar.time = oneWeekAgo
-        val startOfPreviousWeek = calendar.startOfCurrentWeek()
+        currentRange = AnalyticsHubTimeRange(
+            start = calendar.startOfCurrentWeek(),
+            end = calendar.endOfCurrentWeek()
+        )
+
+        val twoWeeksAgo = oneWeekAgo.oneWeekAgo()
+        calendar.time = twoWeeksAgo
         previousRange = AnalyticsHubTimeRange(
-            start = startOfPreviousWeek,
-            end = oneWeekAgo
+            start = calendar.startOfCurrentWeek(),
+            end = calendar.endOfCurrentWeek()
         )
     }
 }
