+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/ranges/data/TodayRangeData.kt
@@ -1,31 +1,32 @@
 package com.woocommerce.android.ui.analytics.ranges.data
 
-import com.woocommerce.android.extensions.oneMonthAgo
-import com.woocommerce.android.extensions.startOfCurrentMonth
+import com.woocommerce.android.extensions.oneDayAgo
+import com.woocommerce.android.extensions.startOfCurrentDay
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRange
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRangeData
 import java.util.Calendar
 import java.util.Date
 
-class AnalyticsHubMonthToDateRangeData(
+class TodayRangeData(
     referenceDate: Date,
     calendar: Calendar
 ) : AnalyticsHubTimeRangeData {
     override val currentRange: AnalyticsHubTimeRange
+
     override val previousRange: AnalyticsHubTimeRange
 
     init {
         calendar.time = referenceDate
         currentRange = AnalyticsHubTimeRange(
-            start = calendar.startOfCurrentMonth(),
+            start = calendar.startOfCurrentDay(),
             end = referenceDate
         )
 
-        val oneMonthAgo = referenceDate.oneMonthAgo()
-        calendar.time = oneMonthAgo
+        val yesterday = referenceDate.oneDayAgo()
+        calendar.time = yesterday
         previousRange = AnalyticsHubTimeRange(
-            start = calendar.startOfCurrentMonth(),
-            end = oneMonthAgo
+            start = calendar.startOfCurrentDay(),
+            end = yesterday
         )
     }
 }
