+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/ranges/data/YesterdayRangeData.kt
@@ -1,5 +1,6 @@
 package com.woocommerce.android.ui.analytics.ranges.data
 
+import com.woocommerce.android.extensions.endOfCurrentDay
 import com.woocommerce.android.extensions.oneDayAgo
 import com.woocommerce.android.extensions.startOfCurrentDay
 import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRange
@@ -7,26 +8,26 @@ import com.woocommerce.android.ui.analytics.ranges.AnalyticsHubTimeRangeData
 import java.util.Calendar
 import java.util.Date
 
-class AnalyticsHubTodayRangeData(
+class YesterdayRangeData(
     referenceDate: Date,
     calendar: Calendar
 ) : AnalyticsHubTimeRangeData {
     override val currentRange: AnalyticsHubTimeRange
-
     override val previousRange: AnalyticsHubTimeRange
 
     init {
-        calendar.time = referenceDate
-        currentRange = AnalyticsHubTimeRange(
-            start = calendar.startOfCurrentDay(),
-            end = referenceDate
-        )
-
         val yesterday = referenceDate.oneDayAgo()
         calendar.time = yesterday
+        currentRange = AnalyticsHubTimeRange(
+            start = calendar.startOfCurrentDay(),
+            end = calendar.endOfCurrentDay()
+        )
+
+        val dayBeforeYesterday = yesterday.oneDayAgo()
+        calendar.time = dayBeforeYesterday
         previousRange = AnalyticsHubTimeRange(
             start = calendar.startOfCurrentDay(),
-            end = yesterday
+            end = calendar.endOfCurrentDay()
         )
     }
 }
